update user schema and add store
