+++ lib/user/resolver.js
@@ -1,42 +1,18 @@
-const model = require('../../model');
-
-const Contact = model.Contact;
-const User = model.User;
+const contactStore = require('../contact/store');
+const userStore = require('./store');
 
 const resolver = {
 	User: {
-		contacts: (_, {contactIds}) => Contact.find({id: contactIds}),
+		contacts: (_, {contactIds}) => contactStore.getContactsByArgs({id: contactIds}),
 	},
 	Query: {
-		user: (_, {id}) => {
-			return User.findOne(id);
-		},
-		users: (_, args) => {
-			const condition = {};
-			if (args.name) condition.name = {$regex: args.name, $options: 'i'};
-			if (args.country) condition.country = args.country;
-			console.log('User conditions: ', condition);
-			let query = User.find(condition);
-			if (args.limit) query = query.limit(args.limit);
-			return query;
-		},
+		user: (_, {id}) => userStore.getUserById(id),
+		users: (_, args) => userStore.getUsersByArgs(args),
 	},
 	Mutation: {
-		createUser: (_, user) => {
-			const newUser = new User(user);
-			return newUser.save();
-		},
-		updateUser: (_, user) => {
-			const {id, ...rest} = user;
-			return User.findByIdAndUpdate(id, {$set: rest}, {new: true})
-				.catch(err => console.error(err));
-		},
-		deleteUser: (_, {id}) => {
-			return User.findByIdAndDelete(id)
-				.then(user => user.remove())
-				.then(() => `${id} successfully deleted`)
-				.catch(err => console.error(err));
-		},
+		createUser: (_, user) => userStore.createUser(user),
+		updateUser: (_, user) => userStore.updateUser(user),
+		deleteUser: (_, {id}) => userStore.deleteUser(id),
 	},
 };
