+++ lib/user/schema.graphql
@@ -17,7 +17,8 @@ extend type Query {
 		id: [String],
 		name: [String],
 		country: String,
-		limit: Int,
+		first: Int,
+		after: String,
 	): [User]
 }
